+++ src/api/ThemeType.d.ts
@@ -1,7 +1,7 @@
 /**
  * Global type for a theme.
  */
-export type StyleType = {
+export type ThemeType = {
   default: {
     white: string;
     black: string;
