+++ src/components/structure/ListItem.tsx
@@ -1,7 +1,7 @@
+import Separator from '@components/structure/Separator';
 import React, { Component, PropsWithChildren } from 'react';
 import { GestureResponderEvent, StyleProp, StyleSheet, TouchableWithoutFeedback, View, ViewStyle } from 'react-native';
 import { v4 as uuid } from 'react-native-uuid';
-import Separator from '@components/structure/Separator';
 
 interface ListItemProps {
   /**
