+++ src/components/structure/ViewBase.tsx
@@ -1,8 +1,7 @@
-import React, { PropsWithChildren } from 'react';
-import { StyleSheet, View } from 'react-native';
-
 import Container from '@components/structure/Container';
 import variables from '@config/variables';
+import React, { PropsWithChildren } from 'react';
+import { StyleSheet, View } from 'react-native';
 
 type ViewBaseProps = {};
