+++ src/context/FlyoutContext.tsx
@@ -1,8 +1,7 @@
-import React, { PropsWithChildren, ReactNode, useState } from 'react';
-import { Text } from 'react-native';
-
 import { FlyoutType } from '@api/FlyoutType';
 import Flyout from '@components/structure/Flyout';
+import React, { PropsWithChildren, ReactNode, useState } from 'react';
+import { Text } from 'react-native';
 
 /**
  * @deprecated use flyout directly instead
