+++ src/context/PescaContext.tsx
@@ -1,8 +1,6 @@
-import React, { PropsWithChildren } from 'react';
-
 import { PescaClient } from '@api/PescaClient';
-
 import createPescaClient from '@pesca/createPescaClient';
+import React, { PropsWithChildren } from 'react';
 
 export const PescaContext = React.createContext<PescaClient | undefined>(undefined);
